+++ datapipe/step/batch_transform.py
@@ -425,6 +425,11 @@ class BaseBatchTransformStep(ComputeStep):
             for tbl in self.input_dts
         ]
 
+        # Filter out ctes that do not have intersection with transform keys.
+        # These ctes do not affect the result, but may dramatically increase
+        # size of the temporary table during query execution.
+        inp_ctes = [(keys, cte) for (keys, cte) in inp_ctes if len(keys) > 0]
+
         inp = _make_agg_of_agg(inp_ctes, "update_ts")
 
         tr_tbl = self.meta_table.sql_table