+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Fix pandas warning #286
 * Fix SQLAlchemy 2.0 warnings
 * Enable SQLAlchemy 2.0
+* Optimize join in batch transform when there's an input without intersecting
+  keys with transform
 
 # 0.13.6